uc3 works even better with smaller group by windows

+++ src/main/java/allezon/domain/AggregatedValue.java
@@ -4,19 +4,19 @@ import lombok.Data;
 
 @Data
 public class AggregatedValue {
-    private int count;
-    private int price;
+    private long count;
+    private long price;
 
     public AggregatedValue() {
         
     }
 
-    public AggregatedValue(int count, int price) {
+    public AggregatedValue(long count, long price) {
         this.count = count;
         this.price = price;
     }
 
-    public AggregatedValue aggregateProduct(int price) {
+    public AggregatedValue aggregateProduct(long price) {
         this.count += 1;
         this.price += price;
         return this;
